+++ Panther/src/sema_to_pir/SemaToPIRData.hpp
@@ -97,6 +97,8 @@ namespace pcit::panther{
 				pir::ExternalFunction::ID print               = pir::ExternalFunction::ID::dummy();
 				pir::ExternalFunction::ID println             = pir::ExternalFunction::ID::dummy();
 				pir::ExternalFunction::ID get_integer_type_id = pir::ExternalFunction::ID::dummy();
+				pir::ExternalFunction::ID alloc               = pir::ExternalFunction::ID::dummy();
+				pir::ExternalFunction::ID dealloc             = pir::ExternalFunction::ID::dummy();
 			};
 
 			using VTableID = SemaToPIRDataVTableID;